Fix circulr dep
