+++ components/graph/graph.js
@@ -2,8 +2,7 @@ const ko = require('knockout');
 const _ = require('lodash');
 const octicons = require('octicons');
 const components = require('ungit-components');
-const { NodeViewModel } = require('./node');
-const { RefViewModel } = require('./git-ref');
+const { NodeViewModel, RefViewModel } = require('./git-elements');
 const { ComponentRoot } = require('../ComponentRoot');
 const numberOfNodesPerLoad = ungit.config.numberOfNodesPerLoad;
 const { NodesEdges } = require('./nodes-edges');
